# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beyond HOA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List residents
      responses:
        "200":
          description: List of residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resident"
    post:
      operationId: createResident
      tags: [residents]
      summary: Create a resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
  /residents/{id}:
    put:
      operationId: updateResident
      tags: [residents]
      summary: Update a resident
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResident
      tags: [residents]
      summary: Delete a resident
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /dues/checkout:
    post:
      operationId: createDuesCheckout
      tags: [dues]
      summary: Create a Stripe checkout session for dues payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuesCheckoutInput"
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuesCheckoutResult"
...
  /dues/payment-status/{sessionId}:
    get:
      operationId: getDuesPaymentStatus
      tags: [dues]
      summary: Get dues payment status by Stripe session id
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuesPayment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/stripe-setup:
    post:
      operationId: setupStripe
      tags: [admin]
      summary: Configure Stripe secret key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeSetupInput"
      responses:
        "200":
          description: Stripe configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeStripeSetup
      tags: [admin]
      summary: Remove Stripe secret key configuration
      responses:
        "200":
          description: Stripe configuration removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    Resident:
      type: object
      required: [id, name, unit, status, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        unit: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        status: { type: string, enum: [owner, tenant] }
        move_in_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        created_at: { type: string, format: date-time }
    ResidentInput:
      type: object
      required: [name, unit, status]
      properties:
        name: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        status: { type: string, enum: [owner, tenant] }
        move_in_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    ResidentUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        status: { type: string, enum: [owner, tenant] }
        move_in_date: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...
    DuesCheckoutInput:
      type: object
      required: [duesId, period, amount]
      properties:
        duesId: { type: string }
        period: { type: string }
        amount: { type: number }
    DuesCheckoutResult:
      type: object
      required: [url, sessionId]
      properties:
        url: { type: string }
        sessionId: { type: string }
    DuesPayment:
      type: object
      required: [id, dues_id, period, amount, status, created_at]
      properties:
        id: { type: integer }
        dues_id: { type: string }
        period: { type: string }
        amount: { type: string }
        stripe_session_id: { type: ["string", "null"] }
        stripe_payment_intent_id: { type: ["string", "null"] }
        status: { type: string, enum: [pending, paid, failed] }
        paid_at: { type: ["string", "null"], format: date-time }
        created_at: { type: string, format: date-time }
...
    StripeSetupInput:
      type: object
      required: [secretKey]
      properties:
        secretKey: { type: string }