# lib/api-zod/src/generated/types/residentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Beyond HOA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentInputStatus } from './residentInputStatus';

export interface ResidentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  status: ResidentInputStatus;
  /** @nullable */
  move_in_date?: string | null;
  /** @nullable */
  notes?: string | null;
}